connect: carve transparency.rs (pure move)
The transparency log leaves main.rs: the RFC 6962 hash/proof
primitives (with their cfg(test) verifiers), the signed tree head and
log keypair, the log read API (sth/entries/proof/consistency/find),
and the org-revocation-list bulletin board whose sightings the log
witnesses, plus the Merkle/STH tests.
